+++ src/main/java/dev/the_fireplace/caterpillar/item/WritablePatternBookItem.java
@@ -1,0 +1,9 @@
+package dev.the_fireplace.caterpillar.item;
+
+import net.minecraft.world.item.WritableBookItem;
+
+public class WritablePatternBookItem extends WritableBookItem {
+    public WritablePatternBookItem(Properties properties) {
+        super(properties);
+    }
+}