+++ src/main/java/dev/the_fireplace/caterpillar/item/WrittenPatternBookItem.java
@@ -1,0 +1,9 @@
+package dev.the_fireplace.caterpillar.item;
+
+import net.minecraft.world.item.WrittenBookItem;
+
+public class WrittenPatternBookItem extends WrittenBookItem {
+    public WrittenPatternBookItem(Properties properties) {
+        super(properties);
+    }
+}