+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.19.4-7.2.2] - 2023-03-15
 
-First release of the forge 1.19.3 version
+First release of the forge 1.19.4 version
 
 ### Fixed