+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+Update inventory if changed while opened
+
 ## [1.19.4-7.2.2] - 2023-03-15
 
 First release of the forge 1.19.4 version